Only actually enqueue scripts if the post uses FancyBox
We can improve the runtime footprint of the plugin by only actually
enqueueing scripts and styles if needed. In particular, if the post at
hand uses no fancybox images, we don't need to do anything at all. I
considered three approaches to implementing this:

* My original idea was to check the post for whether it uses the
[fancybox] shortcode - this idea was quickly discarded once I noticed
that there is no such shortcode :-)

* The next-best idea was to scan the post content for FancyBox markup.
In particular, for HTML elements using the 'fancybox' class. This is
easy to do, but unfortunately requires calling strpos() on the
(potentially very long) post content.

* I also considered doing all checking on the client side, in
JavaScript. I.e. the scripts and stylesheets are not enqueued on the
server - instead, the client tests whether
'document.querySelector(".fancybox")' yields any hits: if so, all
scripts and styles are loaded dynamically. This seemed to require large
rewrites to the plugin though (and it also seems wasteful to do this on
the client).

Hence, I went for the second approach - which is what this commit
attempts to implement.